fix: -Wsign-compare in Orient and MedialAxis, 13 -> 0
Orient.cpp 7 -> 0, Geometry/MedialAxis.cpp 6 -> 0. Running total 79 of 206.

Orient.cpp is fixed at the source rather than at each comparison:
- `int face_count = mesh->facets_count()` becomes size_t. facets_count() is
  already unsigned, the three loops over it are all `size_t i`, and one
  declaration change clears three warnings.
- the `int num_directions = 10` parameter of area_cumulation and
  area_cumulation_accurate becomes size_t. It is a count, both call sites pass
  positive literals (10 and 14), and it makes the existing
  `std::min(static_cast<size_t>(num_directions), align_counts.size())` cast
  redundant, so the cast goes too.
- the best-orientation scan used `i < results_vector.size() - 1`, which
  underflows to SIZE_MAX on an empty vector; it now compares i + 1 < size().

MedialAxis.cpp is all forward index loops that only ever subscript, so they
become size_t: the two crosspoint scans, the two `far_idx` walks (which also
feed points.begin() + far_idx), the min-width pass, and the roundness scan.

NOT fixed, flagged for a decision: the first crosspoint loop
(`nb_better_than_me`) indexes crosspoint[0] in its body where the very next loop
uses crosspoint[i]. The body is therefore loop-invariant and the counter can only
come out as 0 or crosspoint.size(). It looks like a typo for [i], but correcting
it changes which thin-wall branches get pruned, which is a behaviour change and
not mine to make silently. Orient.cpp also still has a -Wsequence-point warning
(undefined behaviour) that wants its own look.

Verified. Both TUs compile with -Wsign-compare at 0 and no other flag's count
moved. Runtime, against a binary built from the pre-campaign commit eae5ede and
sliced with the real FrankenPrinter profile (Creality Ender-3 0.6mm / 0.20mm
NORMAL / Overture ABS Grey): all 7 strict-gate panel configs identical, including
sphere_arc and bunny and the support configs.

thinwalls_arc first reported a 1-layer difference. Five runs of one binary show
that config is BIMODAL - four runs land in one outcome and one in the other, with
the identical fingerprint (1 of 99 layers, 0.000869 mm, Perimeter -1) - and the
pre-campaign run matches four of the five post-change runs exactly. Same story
already established for pins and bedcover. A single before/after pair cannot tell
a mode switch from a regression here; the comparison has to be against a set of
runs.

+++ src/libslic3r/Geometry/MedialAxis.cpp
@@ -648,7 +648,7 @@ MedialAxis::remove_bits(ThickPolylines& pp) const
 
         //check if is smaller or the other ones are not endpoits
         int nb_better_than_me = 0;
-        for (int i = 0; i < crosspoint.size(); i++) {
+        for (size_t i = 0; i < crosspoint.size(); i++) {
             if (!pp[crosspoint[0]].endpoints.second || length <= pp[crosspoint[0]].length())
                 nb_better_than_me++;
         }
@@ -656,7 +656,7 @@ MedialAxis::remove_bits(ThickPolylines& pp) const
 
         //check if the length of the polyline is small vs width of the other lines
         coord_t local_max_width = 0;
-        for (int i = 0; i < crosspoint.size(); i++) {
+        for (size_t i = 0; i < crosspoint.size(); i++) {
             local_max_width = std::max(local_max_width, pp[crosspoint[i]].points_width[0]);
         }
         if (length > coordf_t(local_max_width + this->m_min_width))
@@ -1486,7 +1486,7 @@ MedialAxis::concatenate_small_polylines(ThickPolylines& pp) const
                 polyline.points_width.back() = std::min(polyline.points_width[polyline.points_width.size() - 2], best_candidate->points_width[1]);  // second-to-last point (for interpolating over-extrusion clamp)
             }
             //be far enough
-            int far_idx = 1;
+            size_t far_idx = 1;
             while (far_idx < best_candidate->points.size() && polyline.back().coincides_with_epsilon(best_candidate->points[far_idx]))
                 far_idx++;
             polyline.points.insert(polyline.points.end(), best_candidate->points.begin() + far_idx, best_candidate->points.end());
@@ -1570,7 +1570,7 @@ MedialAxis::concatenate_polylines_with_crossing(ThickPolylines& pp)
                 polyline.points_width.back() = std::min(polyline.points_width[polyline.points_width.size() - 2], best_candidate->points_width[1]);  // second-to-last point (for interpolating over-extrusion clamp)
             }
             //be far enough
-            int far_idx = 1;
+            size_t far_idx = 1;
             while (far_idx < best_candidate->points.size() && polyline.back().coincides_with_epsilon(best_candidate->points[far_idx]))
                 far_idx++;
             polyline.points.insert(polyline.points.end(), best_candidate->points.begin() + far_idx, best_candidate->points.end());
@@ -1846,7 +1846,7 @@ MedialAxis::grow_to_nozzle_diameter(ThickPolylines& pp, const ExPolygons& anchor
         1, false).scaled_width();
     //ensure the width is not lower than min_width.
     for (ThickPolyline& polyline : pp) {
-        for (int i = 0; i < polyline.points.size(); ++i) {
+        for (size_t i = 0; i < polyline.points.size(); ++i) {
             const bool is_anchored = std::any_of(anchors.begin(), anchors.end(),
                 [&polyline, i](const ExPolygon& poly) { return poly.contains(polyline.points[i]); });
             if (!is_anchored && polyline.points_width[i] < min_width)
@@ -1907,7 +1907,7 @@ check_circular(ExPolygon& expolygon, coord_t max_variation) {
         // Computing circle center
         Point center = expolygon.contour.centroid();
         coordf_t radius_min = std::numeric_limits<float>::max(), radius_max = 0;
-        for (int i = 0; i < expolygon.contour.points.size(); ++i) {
+        for (size_t i = 0; i < expolygon.contour.points.size(); ++i) {
             coordf_t dist = expolygon.contour.points[i].distance_to(center);
             radius_min = std::min(radius_min, dist);
             radius_max = std::max(radius_max, dist);
